Added actual end to end flow

+++ paradisewebapp/src/main/scala/name/piotrszul/paradise/domain/Entity.scala
@@ -2,6 +2,6 @@ package name.piotrszul.paradise.domain
 
 import org.codehaus.jackson.annotate.JsonTypeInfo
 
-case class Entity(id:Int, name:String) {
+case class Entity(id:Long, name:String) {
   
 }
